+++ ESFramework/Foundation/NSData+ESExtension.m
@@ -7,7 +7,6 @@
 //
 
 #import "NSData+ESExtension.h"
-#import <Security/SecRandom.h>
 #import <CommonCrypto/CommonDigest.h>
 #import <CommonCrypto/CommonHMAC.h>
 
@@ -45,16 +44,6 @@
     }
 
     return [NSData dataWithBytesNoCopy:bytes length:bytesLength freeWhenDone:YES];
-}
-
-+ (nullable NSData *)randomDataWithLength:(NSUInteger)length
-{
-    NSMutableData *data = [NSMutableData dataWithLength:length];
-    int result = SecRandomCopyBytes(kSecRandomDefault, (size_t)length, data.mutableBytes);
-    if (0 != result) {
-        return nil;
-    }
-    return [data copy];
 }
 
 - (NSString *)UTF8String