+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Moved `-[NSString UUIDString]` to `ESUUIDString()` helper
 - Moved `-[NSString randomStringWithLength:]` to `ESRandomString()` helper
+- Moved `-[NSData randomDataWithLength:]` to `ESRandomData()` helper
 
 ## 3.16.1 (2019-07-25)